pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ares-ultimate"
version = "1.0.0"
description = "ARES-Ultimate-251129: Level 6 Autonomous Finance System with IBKR/KIS Integration"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "ARES Development Team"}
]
keywords = ["algorithmic-trading", "quantitative-finance", "risk-management"]

dependencies = [
    # Core
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "numba>=0.58.0",
    
    # Data Sources
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "polygon-api-client>=1.12.0",
    "yfinance>=0.2.30",
    "fredapi>=0.5.1",
    
    # Brokers
    "ib_insync>=0.9.86",
    
    # ML/Analytics
    "scikit-learn>=1.3.0",
    
    # Infrastructure
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "asyncio-throttle>=1.0.2",
    
    # Scheduling
    "apscheduler>=3.10.0",
    
    # Monitoring API
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.0.290",
    "mypy>=1.5.0",
]

gpu = [
    "torch>=2.1.0",
    "lightgbm>=4.1.0",
    "xgboost>=2.0.0",
]

[project.scripts]
ares-backtest = "backtest.run_backtest:main"
ares-live = "orchestration.live_orchestrator:main"
ares-dashboard = "monitoring.api:run_server"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "data*", "engines*", "brokers*", "ensemble*", "risk*", "orchestration*", "backtest*", "monitoring*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
